+++ ChangeLog.md
@@ -1,13 +1,22 @@
 FuelSDK-Ruby
 ============
 
-2013-09-11: Version 0.1.1  
+2013-09-18: Version 0.1.2
+```
+  get properties for dataextension for retrieve and edit calls
+
+  clear soap client on refresh so the client gets re-established with header with new token
+
+  refresh tests
+```
+
+2013-09-11: Version 0.1.1
 ```
  Added ChangeLog
- 
+
  soap_configure, soap_perform with supporting tests
- 
+
  make soap_cud more rubular and easier to read and support
- 
+
  fixed some issues when trying to make requests after being passed a jwt
 ```